fix: generate noise in float32 instead of float16
Float16 noise has limited precision that creates quantization patterns
visible as vertical ripple artifacts, especially with few-step Lightning
LoRA generation where the denoiser doesn't have enough steps to smooth
them out. Use float32 (matching Z-Image and the diffusers pipeline).

+++ invokeai/app/invocations/qwen_image_denoise.py
@@ -132,7 +132,7 @@ class QwenImageDenoiseInvocation(BaseInvocation, WithMetadata, WithBoard):
         seed: int,
     ) -> torch.Tensor:
         rand_device = "cpu"
-        rand_dtype = torch.float16
+        rand_dtype = torch.float32
 
         return torch.randn(
             batch_size,
